using jpa crud repository, instead of array list dao

+++ backend/src/main/java/me/fairygel/fbook/controller/BookStatusController.java
@@ -14,7 +14,7 @@ public class BookStatusController {
     private final BookStatusService bookStatusService;
 
     @GetMapping(value = {"/{id}", "/{id}/"})
-    public BookStatus read(@PathVariable Long id) {
+    public BookStatus read(@PathVariable Short id) {
         return bookStatusService.read(id);
     }
